aider: Set `hsh` to the hash of the current git repository using the `subprocess` module.
# Aider chat conversation:

USER: do the todo

ASSISTANT: To complete the TODO, we need to set `hsh` to the hash of the current git repository. We can achieve this by using the `subprocess` module to run a git command and capture the output.

Here is the *SEARCH/REPLACE* block to implement this:

harness.py
```python
<<<<<<< SEARCH
    # todo set hsh to the hash of the current git repo
    prefix = "aider-{hsh}-"
=======
    # Get the current git commit hash
    hsh = subprocess.check_output(["git", "rev-parse", "HEAD"]).decode().strip()
    prefix = f"aider-{hsh}-"
>>>>>>> REPLACE
```